requirements.txt
APScheduler==3.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-CacheControl==0.3.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
